+++ bin/www
@@ -1,4 +1,6 @@
 #!/usr/bin/env node
+'use strict';
+
 var debug =   require('debug')('lialu:'+process.pid),
     libs =    process.cwd() + '/libs/',
     config =  require(libs + 'config'),
