+++ generateData.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var faker = require('faker');
 
 var libs = process.cwd() + '/libs/';
