+++ libs/config.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var nconf = require('nconf');
 
 nconf.argv()
