+++ libs/db/mongoose.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var mongoose = require('mongoose');
 
 var libs = process.cwd() + '/libs/';
