+++ libs/log.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var winston = require('winston');
 
 winston.emitErrs = true;
@@ -27,7 +29,7 @@ function logger(module) {
     });
 }
 
-function getFilePath (module ) {
+function getFilePath (module) {
     //using filename in log statements
     return module.filename.split('/').slice(-2).join('/');
 }
