+++ libs/model/client.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var mongoose = require('mongoose'),
 	Schema = mongoose.Schema,
