+++ libs/model/user.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var mongoose = require('mongoose'),
 	crypto = require('crypto'),
