+++ libs/routes/oauth.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var express = require('express');
 
 var libs = process.cwd() + '/libs/';
