+++ libs/routes/users.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var express = require('express');
 var passport = require('passport');
 var router = express.Router();
